requirements.txt
amqp==2.5.0
billiard==3.6.0.0
blessings==1.7
bpython==0.18
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
curtsies==0.3.0
Django==2.2.3
greenlet==0.4.15
idna==2.8
kombu==4.6.3
Pygments==2.4.2
python-masscan==0.1.3
pytz==2019.1
redis==3.3.4
requests==2.22.0
six==1.12.0
Jinja2==2.10.1
sqlparse==0.3.0
typing==3.7.4
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
